Refactor OpenQueueCard to seprate file
