+++ packages/app/pages/today.tsx
@@ -1,9 +1,9 @@
-import { Avatar, Button, Card, Col, PageHeader, Row, Result } from "antd";
+import { Button, Col, PageHeader, Row, Result } from "antd";
 import useSWR from "swr";
 import Schedule from "./schedule";
-import { Queue, QueuePartial } from "../../common/index";
 import { API } from "@template/api-client";
 import styled from "styled-components";
+import OpenQueueCard from "../components/Today/OpenQueueCard";
 
 const Navbar = () => {
   return (
@@ -13,20 +13,6 @@ const Navbar = () => {
   );
 };
 
-type QueueCardProps = {
-  queue: QueuePartial;
-};
-
-const PaddedCard = styled(Card)`
-  margin-bottom: 25px;
-`;
-
-const AvatarContainer = styled.div`
-  padding-left: 25px;
-  padding-right: 25px;
-  float: left;
-`;
-
 const CreateQueueButton = styled(Button)`
   bottom: 0%;
   width: 20%;
@@ -34,27 +20,6 @@ const CreateQueueButton = styled(Button)`
   background-color: #4cbb17;
   color: white;
 `;
-
-const QueueCard = ({ queue }: QueueCardProps) => {
-  const staffList = queue.staffList;
-  return (
-    <PaddedCard
-      title={staffList.map((staffMember) => staffMember.name).join(", ")}
-      extra={
-        <Button type="primary" size={"middle"}>
-          Join Queue
-        </Button>
-      }
-    >
-      <h1>{queue.room}</h1>
-      {staffList.map((staffMember) => (
-        <AvatarContainer key={staffMember.id}>
-          <Avatar size={128} src={staffMember.photoURL} shape="square" />
-        </AvatarContainer>
-      ))}
-    </PaddedCard>
-  );
-};
 
 export default function Today() {
   const { data, error } = useSWR(`api/v1/courses/1/queue`, async () =>
@@ -78,7 +43,7 @@ export default function Today() {
       <Row gutter={25}>
         <Col md={12} sm={24}>
           {data?.map((q) => (
-            <QueueCard key={q.id} queue={q} />
+            <OpenQueueCard key={q.id} queue={q} />
           ))}
           {isTA ? (
             <CreateQueueButton type="default" size={"large"}>
